Now datatypes type number converted to integer

+++ src/test/files/MS3-to-OAS-20/ms3-datatypes-to-oas.ts
@@ -185,7 +185,7 @@ export const oasDataTypes: OASInterface.API = {
           'required': true
         },
         'isNumber': {
-          'type': 'long'
+          'type': 'integer'
         },
         'isDateOnly': {
           'type': 'date'
