+++ src/test/ms3-datatypes-to-oas.test.ts
@@ -198,7 +198,7 @@ test('MS3 schemas should be converted to OAS successfully', async() => {
               'required': true
             },
             'isNumber': {
-              'type': 'long'
+              'type': 'integer'
             },
             'isDateOnly': {
               'type': 'date'
